Move Slack to collectorlib
